+++ lib/grpc_mock/action_stub.rb
@@ -1,24 +1,71 @@
 # frozen_string_literal: true
 
+require 'active_support/core_ext/module'
+require 'active_support/core_ext/object'
+
 module GrpcMock
   class ActionStub < RequestStub
-    # @param path [String] gRPC path like /${service_name}/${method_name}
-    # @param rpc_action [GRPC::RpcDesc] instance with parameters like +name+, +input+, +output+ and others
-    def initialize(path, rpc_action)
-      @rpc_action = rpc_action
+    class InvalidProtoType < StandardError; end
+
+    attr_reader :service, :action
+
+    # @param service [GRPC::GenericService] gRPC service class representing the the service being stubbed
+    # @param action [String, Symbol] name of the endpoint being stubbed
+    def initialize(service, action)
+      @service = service
+      @action = action
+
       super(path)
     end
 
     # @param proto [Object] request proto object
     # @param request [Hash] parameters for request
     def with(proto = nil, **request)
-      proto ? super(proto) : super(@rpc_action.input.new(request))
+      return super(input_type.new(**request)) if proto.blank?
+      return super(input_type.new(**proto)) if proto.respond_to?(:to_hash)
+
+      raise InvalidProtoType, "Invalid proto type #{proto.class} for #{path}, expected #{input_type}" unless proto.class == input_type
+
+      super(proto)
     end
 
     # @param proto [Object] response proto object
     # @param response [Hash] parameters to respond with
     def to_return(proto = nil, **response)
-      proto ? super(proto) : super(@rpc_action.output.new(response))
+      return super(output_type.new(**response)) if proto.blank?
+      return super(output_type.new(**proto)) if proto.respond_to?(:to_hash)
+
+      raise InvalidProtoType, "Invalid proto type #{proto.class} for #{path}, expected #{output_type}" unless proto.class == output_type
+
+      super(proto)
+    end
+
+    private
+
+    delegate :service_name, :rpc_descs, to: :grpc_service
+
+    def grpc_service
+      service.const_defined?(:Service) ? service::Service : service
+    end
+
+    def endpoint_name
+      @endpoint_name ||= rpc_descs.key?(action) ? action : action.to_s.camelize.to_sym
+    end
+
+    def rpc_desc
+      rpc_descs[endpoint_name]
+    end
+
+    def path
+      "/#{service_name}/#{endpoint_name}"
+    end
+
+    def input_type
+      rpc_desc.input
+    end
+
+    def output_type
+      rpc_desc.output
     end
   end
 end