+++ spec/grpc_mock/action_stub_spec.rb
@@ -3,15 +3,17 @@
 require 'examples/hello/hello_services_pb'
 
 RSpec.describe GrpcMock::ActionStub do
-  subject(:action_stub) { described_class.new(path, action) }
+  subject(:action_stub) { described_class.new(service, endpoint) }
 
-  let(:path) do
-    '/service_name/method_name'
-  end
+  let(:path) { "/#{service_name}/#{endpoint.to_s.camelize}" }
 
   let(:action) do
     double(input: input_class, output: output_class)
   end
+
+  let(:service) { Hello::Hello::Service }
+  let(:service_name) { service.service_name  }
+  let(:endpoint) { :hello }
 
   let(:input_class) { Hello::HelloRequest }
   let(:output_class) { Hello::HelloResponse }
@@ -55,51 +57,87 @@ RSpec.describe GrpcMock::ActionStub do
   end
 
   describe '#with' do
-    context "when hash params" do
-      let(:request) { { msg: "request" } }
+    before { allow(input_class).to receive(:new).and_call_original }
+
+    context 'with a hash' do
+      let(:request) { { msg: 'request' } }
 
       it 'registers request', aggregate: true do
-        expect(action).to receive(:input)
-        expect(input_class).to receive(:new).with(request).and_call_original
 
         expect(action_stub.with(request)).to eq(action_stub)
+        expect(input_class).to have_received(:new).with(request)
       end
     end
 
-    context "when proto object" do
-      let(:request) { :request }
+    context 'with kwargs' do
+      let(:request) { { msg: 'Hello?' } }
 
       it 'registers request', aggregate: true do
-        expect(action).not_to receive(:input)
-        expect(input_class).not_to receive(:new).with(request)
+        expect(action_stub.with(**request)).to eq(action_stub)
+        expect(input_class).to have_received(:new).with(request)
+      end
+    end
 
+    context 'with a proto object' do
+      let(:request) { input_class.new(msg: 'hello?') }
+
+      it 'registers request', aggregate: true do
         expect(action_stub.with(request)).to eq(action_stub)
+      end
+
+      context 'with wrong proto class' do
+        let(:request) { output_class.new(msg: 'hello?') }
+
+        it 'raises InvalidProtoType' do
+          expect { expect(action_stub.with(request)) }.to raise_error described_class::InvalidProtoType
+        end
       end
     end
   end
 
   describe '#to_return' do
-    context "when hash params" do
-      let(:response) { { msg: "response" } }
+    before { allow(GrpcMock::ResponsesSequence).to receive(:new).and_call_original }
+
+    context 'with a hash' do
+      let(:response) { { msg: 'Hello!' } }
 
       it 'registers response', aggregate: true do
-        expect(action).to receive(:output)
         expect(output_class).to receive(:new).with(response)
 
-        expect(GrpcMock::ResponsesSequence).to receive(:new).with([GrpcMock::Response::Value]).once
         expect(action_stub.to_return(response)).to eq(action_stub)
+
+        expect(GrpcMock::ResponsesSequence).to have_received(:new).with([GrpcMock::Response::Value]).once
       end
     end
 
-    context "when proto object" do
-      let(:response) { :response }
+    context 'with kwargs' do
+      let(:response) { { msg: 'Hello!' } }
+      it 'registers response', aggregate: true do
+        expect(output_class).to receive(:new).with(**response)
+
+        expect(action_stub.to_return(**response)).to eq(action_stub)
+
+        expect(GrpcMock::ResponsesSequence).to have_received(:new).with([GrpcMock::Response::Value]).once
+      end
+    end
+
+    context 'with a proto object' do
+      let(:response) { output_class.new(msg: 'Hello!') }
 
       it 'registers response', aggregate: true do
-        expect(action).not_to receive(:output)
         expect(output_class).not_to receive(:new).with(response)
 
-        expect(GrpcMock::ResponsesSequence).to receive(:new).with([GrpcMock::Response::Value]).once
         expect(action_stub.to_return(response)).to eq(action_stub)
+
+        expect(GrpcMock::ResponsesSequence).to have_received(:new).with([GrpcMock::Response::Value]).once
+      end
+
+      context 'with wrong proto class' do
+        let(:response) { input_class.new(msg: 'hello?') }
+
+        it 'raises InvalidProtoType' do
+          expect { expect(action_stub.to_return(response)) }.to raise_error described_class::InvalidProtoType
+        end
       end
     end
   end
@@ -146,6 +184,8 @@ RSpec.describe GrpcMock::ActionStub do
   end
 
   describe '#match?' do
-    it { expect(action_stub.match?(path, double(:request))).to eq(true) }
+    subject { action_stub.match?(path, double) }
+
+    it { is_expected.to be true }
   end
 end