+++ CHANGELOG.md
@@ -5,5 +5,5 @@
 - Rename gem to `gruffish`
 - Add support for `return_op` option([#1](https://github.com/Freshly/gruffish/pull/1), [#2](https://github.com/Freshly/gruffish/pull/2))
 - Fix load order bug ([#3](https://github.com/Freshly/gruffish/pull/3))
-- Add `stub_grpc_action` convenience method ([#9](https://github.com/Freshly/gruffish/pull/9))
+- Add `stub_grpc_action` convenience method ([#9](https://github.com/Freshly/gruffish/pull/9), [#12](https://github.com/Freshly/gruffish/pull/12))
 - Add DSL module for RSpec ([#10](https://github.com/Freshly/gruffish/pull/10))